fix: refactor network
